(#7621) Windows OS Release Fact and windows support for osrelease
This patch queries wmi for the Caption property in the
Win32_OperatingSystem class. The result is a string that
is parsed by comparing it to a hardcode list of possible
versions. A hardcoded list is not ideal but Microsoft does
not provide a function to grab just the release string.

A possible alternative would be using regex to grab the
value, accounting for the various cases between windows
version names and then hoping Mircrosoft doesn't change
the format of the caption property in future releases.

The patch also adds a windowsrelease fact which is the full
name of the specific version of windows.
ie. "Microsoft Windows 7 Professional"

Signed-off-by: William Van Hevelingen <wvan13@gmail.com>